move models and remove reports directory
