--- qa/qa/reports/models/__init__.py
@@ -1,4 +1,0 @@
-"""QA report models."""
-
-from qa.reports.models.base import Flag  # noqa
-from qa.reports.models.fulltext import FulltextReport  # noqa
